Another fix for ByteFlow

+++ CyanAPI-1.1-1.17.1/src/main/java/org/asf/cyan/api/internal/modkit/components/_1_17/common/network/buffer/FriendlyByteBufInputFlow.java
@@ -7,7 +7,7 @@ public class FriendlyByteBufInputFlow implements ByteFlow {
 
 	private FriendlyByteBuf buffer;
 	
-	private int next = -1;
+	private byte next = -1;
 	private boolean end = false;
 	private boolean hasNext = false;
 
@@ -16,9 +16,9 @@ public class FriendlyByteBufInputFlow implements ByteFlow {
 	}
 
 	@Override
-	public int read() {
+	public byte read() {
 		if (hasNext) {
-			int i = next;
+			byte i = next;
 			next = -1;
 			hasNext = false;
 			return i;
