+++ ModKit-1.0/src/main/java/org/asf/cyan/api/network/ByteFlow.java
@@ -10,5 +10,5 @@ package org.asf.cyan.api.network;
 public interface ByteFlow {
 	public boolean hasNext();
 
-	public int read();
+	public byte read();
 }
