+++ ModKit-1.2/src/main/java/modkit/network/ByteFlow.java
@@ -10,5 +10,5 @@ package modkit.network;
 public interface ByteFlow {
 	public boolean hasNext();
 
-	public int read();
+	public byte read();
 }
